login en frontend terminado

+++ server/Models/TareasModel.js
@@ -1,7 +1,9 @@
 const sql = require('mssql');
 const dbConfig = require('../Config/dbConfig'); // Configuración de la base de datos
 const { Response, TypeOfResponse } = require('../Common/Response');
-let result = {};
+const jwt = require('jsonwebtoken');
+require('dotenv').config();
+
 
 class TareasModel{
     constructor(id, user_id, title, description, status){
@@ -12,11 +14,17 @@ class TareasModel{
         this.status = status;
     }
 
-    static async GetTasks(){
+    static async GetTasks(token){
         const response = new Response(); // Crear una instancia de Response
         try {
+                // Verificar y decodificar el token
+            const decoded = jwt.verify(token, process.env.JWT_SECRET);
+            const user_id = decoded.id; 
+            
             const pool = await sql.connect(dbConfig);
-            const result = await pool.request().query('SELECT * FROM TASKS');
+            const result = await pool.request()
+            .input('user_id', sql.Int, user_id)
+            .query('SELECT * FROM TASKS WHERE user_id = @user_id');
             
             response.type_of_response = TypeOfResponse.SUCCESS;
             response.message = 'Tareas obtenidas correctamente';
@@ -27,17 +35,22 @@ class TareasModel{
             response.type_of_response = TypeOfResponse.ERROR;
         }
     return response; 
-}
+    }
 
-    static async GetTaskById(id) {
-        const response = new Response();
+
+    static async GetTaskById(id, token) {
         try {
+            console.log("tentra model task")
+            const decoded = jwt.verify(token, process.env.JWT_SECRET);
+            const user_id = decoded.id; 
+            
             // Conectar a la base de datos
             const pool = await sql.connect(dbConfig);
             // Hacer la consulta usando el id
             const result = await pool.request()
                 .input('id', sql.Int, id)
-                .query('SELECT * FROM TASKS WHERE id = @id');
+                .input('user_id', sql.Int, user_id)
+                .query('SELECT * FROM TASKS WHERE id = @id AND user_id = @user_id');
             result.recordset[0].id = id;
             if (result.recordset.length > 0) {
                 return new Response(TypeOfResponse.SUCCESS, result.recordset[0], 'Tarea obtenida correctamente.');
@@ -52,9 +65,11 @@ class TareasModel{
 
     }
 
-    static async AddTask(user_id, title, description, status) {
+    static async AddTask(token, title, description, status) {
         const response = new Response();
         try {   
+            const decoded = jwt.verify(token, process.env.JWT_SECRET);
+            const user_id = decoded.id;
             const pool = await sql.connect(dbConfig);
             const result = await pool.request()
                 .input('user_id', sql.Int, user_id)
@@ -82,18 +97,17 @@ class TareasModel{
         }
     }
 
-    static async DeleteTask(id) {
-        const response = new Response();
-        try {
+    static async DeleteTask(token, id) {
+        try {                
+            const decoded = jwt.verify(token, process.env.JWT_SECRET);
+            const user_id = decoded.id;
             const pool = await sql.connect(dbConfig);
-            const result = await pool.request()
+            let result = await pool.request()
                 .input('id', sql.Int, id)
-                .query('DELETE FROM TASKS WHERE id = @id');
+                .input('user_id', sql.Int, user_id)
+                .query('DELETE FROM TASKS WHERE id = @id AND user_id = @user_id');
             if (result.rowsAffected[0] === 0) {
-                response.type_of_response = TypeOfResponse.WARNING;
-                response.message = 'No se eliminó la tarea';
-                response.data = {};
-                return new Response(TypeOfResponse.WARNING, {}, 'No se eliminó la tarea.');
+                return new Response(TypeOfResponse.WARNING, {}, 'No se encontró la tarea.');
             }
             return new Response(TypeOfResponse.SUCCESS, {}, 'Tarea eliminada correctamente.');
         } catch (error) {
@@ -101,34 +115,25 @@ class TareasModel{
         }
     }
 
-    static async UpdateTask(id, title, description, status) {
-        let response = new Response();
+    static async UpdateTask(token, id, title, description, status) {
         try {
-            console.log("id", id, "title", title, "description", description, "status", status)
-            let token = sessionStorage.getItem('token')
-            console.log("token", token) 
+            const decoded = jwt.verify(token, process.env.JWT_SECRET);
+            const user_id = decoded.id;
             const pool = await sql.connect(dbConfig);
-            console.log("pool", pool)
-            const updateResponse = await pool.request()
-            .input('id', sql.Int, id)
-            .input('title', sql.NVarChar(255), title)
-            .input('description', sql.NVarChar(sql.MAX), description)
-            .input('status', sql.NVarChar(50), status)
-            .query(`
-                UPDATE TASKS
-                SET title = @title, description = @description, status = @status
-                WHERE id = @id 
-                OUTPUT inserted.*
-                SELECT * FROM TASKS WHERE id = @id;
-            `);
-    
-        console.log("Updated Task:", updateResponse.recordset); 
-                console.log(response)
-            if (response.rowsAffected[0] !== 0) {
-                return new Response(TypeOfResponse.SUCCESS, result.recordset[0], 'Tarea actualizada correctamente.');
-            } else {
-                return new Response(TypeOfResponse.ERROR, {}, 'Error al actualizar la tarea.');
-            }
+            let result = await pool.request()
+                .input('id', sql.Int, id)
+                .input('user_id', sql.Int, user_id)
+                .input('title', sql.NVarChar, title)
+                .input('description', sql.NVarChar, description)
+                .input('status', sql.NVarChar, status)
+                .query(`
+                    UPDATE TASKS SET title = @title, description = @description, status = @status
+                    WHERE id = @id AND user_id = @user_id
+                `);
+                if (result.rowsAffected[0] !== 0) {
+                console.log("actualizada con exito")
+                return new Response(TypeOfResponse.SUCCESS, {}, 'Tarea actualizada correctamente.');
+                }
         } catch (error) {
             return new Response(TypeOfResponse.ERROR, {}, 'Error al conectar al servidor.');
         }
